+++ canon/src/store/host.rs
@@ -5,18 +5,16 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use blake2b_simd::Params;
-use lazy_static::lazy_static;
-use parking_lot::RwLock;
 
+use std::cell::RefCell;
 use std::collections::HashMap;
-use std::sync::Arc;
 
-use crate::id::IdHash;
-use crate::CanonError;
+use crate::canon::CanonError;
+use crate::id::{Id, IdHash};
 
-lazy_static! {
-    static ref STATIC_MAP: Arc<RwLock<HashMap<IdHash, Vec<u8>>>> =
-        Arc::new(RwLock::new(HashMap::new()));
+thread_local! {
+    pub static STATIC_MAP: RefCell<HashMap<IdHash, Vec<u8>>> =
+    RefCell::new(HashMap::new())
 }
 
 pub(crate) struct HostStore;
@@ -26,11 +24,9 @@ impl HostStore {
         hash: &IdHash,
         into: &mut [u8],
     ) -> Result<(), CanonError> {
-        match STATIC_MAP
-            .read()
-            .get(hash)
-            .map(|vec| into.copy_from_slice(&vec))
-        {
+        match STATIC_MAP.with(|m| {
+            m.borrow().get(hash).map(|vec| into.copy_from_slice(&vec))
+        }) {
             Some(()) => Ok(()),
             None => Err(CanonError::NotFound),
         }
@@ -39,16 +35,24 @@ impl HostStore {
     pub(crate) fn put(bytes: &[u8]) -> IdHash {
         debug_assert!(bytes.len() > core::mem::size_of::<IdHash>());
         let hash = Self::hash(bytes);
-        STATIC_MAP.write().insert(hash, Vec::from(bytes));
+        STATIC_MAP.with(|m| m.borrow_mut().insert(hash, Vec::from(bytes)));
         hash
     }
 
-    pub fn hash(bytes: &[u8]) -> IdHash {
+    pub(crate) fn hash(bytes: &[u8]) -> IdHash {
         let mut state = Params::new().hash_length(32).to_state();
         state.update(bytes);
 
         let mut buf = [0u8; 32];
         buf.copy_from_slice(state.finalize().as_ref());
         buf
+    }
+
+    pub(crate) fn take_bytes(id: &Id) -> Result<Vec<u8>, CanonError> {
+        STATIC_MAP.with(|m| match m.borrow_mut().remove(&id.hash()) {
+            Some(vec) if id.size() == vec.len() => Ok(vec),
+            Some(_) => Err(CanonError::InvalidEncoding),
+            None => Err(CanonError::NotFound),
+        })
     }
 }